+++ crates/matrix-sdk-crypto/CHANGELOG.md
@@ -12,6 +12,10 @@
   keys whose metadata lacked check fields.
   ([#3046](https://github.com/matrix-org/matrix-rust-sdk/pull/3046))
 
+- Add new API `Device::encrypt_event_raw` that allows
+  to encrypt an event to a specific device.
+  ([#3091](https://github.com/matrix-org/matrix-rust-sdk/pull/3091))
+
 # 0.7.0
 
 - Add method to mark a list of inbound group sessions as backed up: